fix: Index handleChecksumFile ignores path to file
When the checksum is stored in a file named identically to the download
file but with a suffix, we kept the path to the file in the information
extracted. In production, the path should not be present, but for
consistency we now explicitly drop it, which is also beneficial for the
tests.