fix: remove sleep in checksums collection
We have a long sleep after downloading checksums, and a timeout for
message retrieval from the queue.
If we download a checksum, we have a long sleep. If we don't, we should
go to the next msg immediately. If the queue is empty, there will be a
timeout of 1s, so we check the queue once every second at most in that
case.