Revert "feat: use an OrderedConsumer looking at mesg since 24h"
This reverts commit de0dcf2b8e9934d84bb78be58e0bc5c3bb0956a1.

It seems the ordered consumer is just meant to peak inside the stream.
It is a push consumer, and I use a pull consumer, and this led to
the error that my consumer requires an ack policy.

I discovered that the pull consumer as I used before also accepts a
DeliverPolicy option, which I will try to use.